requirements.txt
# Core Dependencies
fastapi==0.115.5
uvicorn[standard]==0.32.1
pydantic==2.10.3
pydantic-settings==2.6.1
python-multipart==0.0.12

# Machine Learning & Data Processing
scikit-learn==1.5.2
numpy==1.26.4
pandas==2.2.3
scipy==1.14.1

# NLP & Pattern Matching
spacy==3.8.2
textstat==0.7.3

# Log Parsing
python-dateutil==2.9.0.post0
parse==1.20.2
regex==2024.11.6

# Database & Caching
redis==5.2.0
motor==3.6.0
pymongo>=4.9,<4.10

# Monitoring & Metrics
prometheus-client==0.21.0

# Visualization & Export
plotly==5.24.1
jinja2==3.1.4

# Utils
python-dotenv==1.0.1
httpx==0.27.2
loguru==0.7.2
pyyaml==6.0.2
psutil==6.1.0

# Development Tools
pytest==8.3.3
pytest-asyncio==0.24.0
pytest-cov==6.0.0
black==24.10.0
ruff==0.8.0
mypy==1.13.0
pre-commit==4.0.1